Move switch skill effect to a subdirectory.
