+++ src/app/ffbe/model/effects/abilities/skill/ability-skill-switch-effect.model.spec.ts
@@ -1,8 +1,8 @@
-import {ABILITY_SKILLS_NAMES_TEST_DATA, ABILITY_SKILLS_SHORTDESCRIPTIONS_TEST_DATA, ABILITY_SKILLS_TEST_DATA} from '../../skill.model.spec';
-import {Skill} from '../../skill.model';
-import {SkillsService} from '../../../services/skills.service';
-import {SkillsServiceMock} from '../../../services/skills.service.spec';
-import {AbilitySkillEffectFactory} from '../ability-skill-effect.factory';
+import {ABILITY_SKILLS_NAMES_TEST_DATA, ABILITY_SKILLS_SHORTDESCRIPTIONS_TEST_DATA, ABILITY_SKILLS_TEST_DATA} from '../../../skill.model.spec';
+import {Skill} from '../../../skill.model';
+import {SkillsService} from '../../../../services/skills.service';
+import {SkillsServiceMock} from '../../../../services/skills.service.spec';
+import {AbilitySkillEffectFactory} from '../../ability-skill-effect.factory';
 
 describe('AbilitySkillSwitchEffect', () => {
   it('should parse skill switch after single skill', () => {
